backend/requirements.txt
# Flask核心
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# 数据库
chromadb==0.4.18

# 机器学习
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
numpy==1.24.3

# 文本处理
spacy==3.7.2

# 爬虫
selenium==4.15.2
webdriver-manager==4.0.1

# HTTP请求
requests==2.31.0

# 环境变量
python-dotenv==1.0.0

# 认证和安全
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.7

# 日志
python-json-logger==2.0.7

# 云服务SDK（阿里云）
alibabacloud-tea-openapi==0.3.9
alibabacloud-dashscope==1.14.0

# 内网穿透
pyngrok==7.0.0

# 工具库
flask-limiter==3.6.0
flask-compress==1.14

# 缓存
redis==5.0.1

# API文档
flasgger==0.9.7.1
openapi-spec-validator==0.5.7
